`build-stitching.yml`: Merge workflow steps into single step
This is just temporary, and will save time in debugging. This is because environment variables don't usually carry over between steps. So, we won't have to do any sort of `echo "export THIRD_PARTY_DIRECTORY=3rdParty" >> $GITHUB_ENV` commands if we keep it all in one step for now.

See also: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-environment-variable